# docs/swagger.json
{
    "components": {
        "schemas": {
            "api.ErrorResponse": {
                "properties": {
                    "error": {
                        "example": "validation failed",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
    "info": {
        "description": "Central coordination controller for the atropos ecosystem.\nManages rules, faults, experiments, workflows, and SDK lifecycle.",
        "title": "Manteion Control-Plane API",
        "version": "1.0"
    },
    "paths": {
...
        "/zeus/attacks": {
            "post": {
                "description": "Proxies the request to zeus-go's load generation API.\nBody and response shapes are defined by zeus-go; see its OpenAPI spec.\nThe response status code is propagated from upstream.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        },
                        "description": "passthrough \u2014 see zeus-go spec for response shape"
                    },
                    "502": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "zeus unreachable"
                    },
                    "default": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "upstream-propagated 4xx/5xx"
                    }
                },
                "summary": "Create attack (Zeus passthrough)",
                "tags": [
                    "zeus-proxy"
                ]
            }
        },
        "/zeus/attacks/{id}": {
            "delete": {
                "description": "Proxies the request to zeus-go's load generation API.",
                "parameters": [
                    {
                        "description": "Attack ID (zeus-defined)",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "passthrough \u2014 see zeus-go spec for response shape"
                    },
                    "502": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "zeus unreachable"
                    },
                    "default": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "upstream-propagated 4xx/5xx"
                    }
                },
                "summary": "Delete attack (Zeus passthrough)",
                "tags": [
                    "zeus-proxy"
                ]
            },
            "get": {
                "description": "Proxies the request to zeus-go's load generation API.\nResponse shape is defined by zeus-go; see its OpenAPI spec.",
                "parameters": [
                    {
                        "description": "Attack ID (zeus-defined)",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        },
                        "description": "passthrough \u2014 see zeus-go spec for response shape"
                    },
                    "502": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "zeus unreachable"
                    },
                    "default": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "upstream-propagated 4xx/5xx"
                    }
                },
                "summary": "Get attack (Zeus passthrough)",
                "tags": [
                    "zeus-proxy"
                ]
            }
        },
        "/zeus/workloads": {
            "get": {
                "description": "Proxies the request to zeus-go's load generation API.\nResponse shape is defined by zeus-go; see its OpenAPI spec.",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        },
                        "description": "passthrough \u2014 see zeus-go spec for response shape"
                    },
                    "502": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "zeus unreachable"
                    },
                    "default": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "upstream-propagated 4xx/5xx"
                    }
                },
                "summary": "List workloads (Zeus passthrough)",
                "tags": [
                    "zeus-proxy"
                ]
            },
            "post": {
                "description": "Proxies the request to zeus-go's load generation API.\nBody and response shapes are defined by zeus-go; see its OpenAPI spec.\nThe response status code is propagated from upstream.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        },
                        "description": "passthrough \u2014 see zeus-go spec for response shape"
                    },
                    "502": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "zeus unreachable"
                    },
                    "default": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "upstream-propagated 4xx/5xx"
                    }
                },
                "summary": "Create workload (Zeus passthrough)",
                "tags": [
                    "zeus-proxy"
                ]
            }
        },
        "/zeus/workloads/{id}": {
            "delete": {
                "description": "Proxies the request to zeus-go's load generation API.",
                "parameters": [
                    {
                        "description": "Workload ID (zeus-defined)",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "passthrough \u2014 see zeus-go spec for response shape"
                    },
                    "502": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "zeus unreachable"
                    },
                    "default": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "upstream-propagated 4xx/5xx"
                    }
                },
                "summary": "Delete workload (Zeus passthrough)",
                "tags": [
                    "zeus-proxy"
                ]
            }
        }
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "description": "Local dev (HTTP)",
            "url": "http://localhost:8080/api/v1"
        },
        {
            "description": "Local dev (HTTPS)",
            "url": "https://localhost:8080/api/v1"
        }
    ]
}